requirements.txt
absl-py==0.9.0
astor==0.8.1
Authlib==0.14.1
boto3==1.12.6
botocore==1.15.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
docutils==0.15.2
Flask==1.1.1
gast==0.3.3
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
python-dateutil==2.8.1
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.11.2
